Changed versioning to pbr, resolved circular input errors, moved base… (#52)
* Changed versioning to pbr, resolved circular input errors, moved base elements to __init__

* fixed import

* Added missing pbr dependency

* Fixed relative import

* Removed circular import

* Removed short description

* Added pbr requirement

* Updated GH-Actions

* Fixed inferred versioning from metadata

* Added description file content type

* enabled build on pr and removed generation of version number

* Changed keywords, updated gh actions

* pypi publish action update
