+++ .github/workflows/python-test_and_deploy.yml
@@ -21,9 +21,9 @@ jobs:
           - 27017:27017
 
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
     - name: Set up Python 3.10.8
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v4
       with:
         python-version: 3.10.8
 #    - name: Install os dependencies
@@ -32,7 +32,7 @@ jobs:
 #        sudo apt-get -y install gfortran swig
     - name: Install dependencies
       run: |
-        pip install wheel flake8 
+        pip install wheel flake8 pbr
         python setup.py egg_info
         pip install tensorflow pytest pytest-cov coveralls -r photonai.egg-info/requires.txt -r photonai/optimization/smac/requirements.txt -r photonai/optimization/nevergrad/requirements.txt
     - name: Test with pytest
@@ -47,23 +47,20 @@ jobs:
     name: Build and publish to TestPyPI
     runs-on: ubuntu-22.04
     needs: pytest
-    if: github.ref == 'refs/heads/master'
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v3
+      with:
+        fetch-depth: 0
     - name: Set up Python 3.10.8
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v4
       with:
         python-version: 3.10.8
-    - name: Replace version number with commit hash
-      run: |
-        dt=$(date '+%Y.%m.%d.%H.%M.%S')
-        sed -i "s/^__version__.*/__version__ = '$dt'/g" setup.py
     - name: Install pypa/build
-      run: pip install build
+      run: pip install build pbr wheel
     - name: Build a binary wheel and a source tarball
-      run: python -m build --sdist --wheel --outdir dist/ .
+      run: python -m build -n --sdist --wheel --outdir dist/ .
     - name: Publish distribution to Test PyPI
-      uses: pypa/gh-action-pypi-publish@master
+      uses: pypa/gh-action-pypi-publish@release/v1
       with:
         user: __token__
         password: ${{ secrets.TEST_PYPI_API_TOKEN }}
