+++ .gitignore
@@ -181,3 +181,6 @@ tmtagsHistory
 .bundle
 /hyperpipe_results.csv
 /config_history.csv
+
+AUTHORS
+ChangeLog
