+++ photonai/__init__.py
@@ -12,7 +12,8 @@ After PHOTON has found the best configuration for your model, it offers a conven
 It also enables you to persist and load your optimal model, including all preprocessing elements, with only one line of code.
 
 """
+from .version import __version__
 
-__version__ = '2.2.1'
+from .base import Hyperpipe, OutputSettings, Stack, Switch, Branch, PipelineElement, ParallelBranch, PhotonRegistry
 
-# __all__ = ( )
+
