+++ photonai/base/hyperpipe.py
@@ -23,7 +23,7 @@ from sklearn.dummy import DummyClassifier, DummyRegressor
 import joblib
 from sklearn.model_selection._split import BaseCrossValidator, BaseShuffleSplit, _RepeatedSplits
 from sklearn.inspection import permutation_importance
-from photonai.__init__ import __version__
+from photonai.version import __version__
 from photonai.base.cache_manager import CacheManager
 from photonai.base.photon_elements import Stack, Switch, Preprocessing, CallbackElement, Branch, PipelineElement, \
     PhotonNative
