+++ requirements.txt
@@ -1,0 +1,17 @@
+numpy
+matplotlib
+scikit-learn
+pandas
+plotly
+imbalanced-learn
+pymodm
+scipy
+statsmodels
+prettytable
+seaborn
+joblib
+dask>=2021.10.0
+distributed
+scikit-optimize
+xlrd
+pbr
