+++ setup.cfg
@@ -1,2 +1,30 @@
 [metadata]
+name = photonai
+description_file = README.md
+long_description_content_type = text/markdown
+license = GPLv3
+author = PHOTONAI Team
+author_email = hahnt@wwu.de
+url = https://www.photon-ai.com/
+project_urls=
+    Source Code = https://github.com/wwu-mmll/photonai/
+    Documentation = https://wwu-mmll.github.io/photonai/
+    Bug Tracker = https://github.com/wwu-mmll/photonai/issues
+
+download_url=https://pypi.org/project/photonai/#files
+keywords=
+    machine learning
+    pipeline
+    research
+    hyperparameter
+    hyperparameter optimization
+classifier=
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: GNU General Public License v3 (GPLv3)
+    Operating System :: OS Independent
+    Intended Audience :: Science/Research
+    Intended Audience :: Developers
 decription-file = README.md
+
+[files]
+packages=photonai
